Cloud Functions/Allow Player/requirements.txt
functions-framework==3.*
google-cloud-compute==1.33.0
